+++ postfix/local/command.c
@@ -102,7 +102,7 @@ int     deliver_command(LOCAL_STATE state, USER_ATTR usr_attr, char *command)
      * 
      * Skip this command if it was already delivered to as this user.
      */
-    if (been_here(state.dup_filter, "command %d %s", usr_attr.uid, command))
+    if (been_here(state.dup_filter, "command %ld %s", (long) usr_attr.uid, command))
 	return (0);
 
     /*
