+++ postfix/pipe/pipe.c
@@ -552,17 +552,17 @@ static void get_service_attr(PIPE_ATTR *attr, char **argv)
     if (attr->uid == var_owner_uid)
 	msg_fatal("request to deliver as mail system owner");
     if (attr->gid == 0)
-	msg_fatal("request to use privileged group id %d", attr->gid);
+	msg_fatal("request to use privileged group id %ld", (long) attr->gid);
     if (attr->gid == var_owner_gid)
-	msg_fatal("request to use mail system owner group id %d", attr->gid);
+	msg_fatal("request to use mail system owner group id %ld", (long) attr->gid);
 
     /*
      * Give the poor tester a clue of what is going on.
      */
     if (msg_verbose)
-	msg_info("%s: uid %d, gid %d, flags %d, size %ld",
-		 myname, attr->uid, attr->gid, attr->flags,
-		 (long) attr->size_limit);
+	msg_info("%s: uid %ld, gid %ld, flags %d, size %ld",
+		 myname, (long) attr->uid, (long) attr->gid,
+		 attr->flags, (long) attr->size_limit);
 }
 
 /* eval_command_status - do something with command completion status */
